+++ truffle/src/com.oracle.truffle.dsl.processor/src/com/oracle/truffle/dsl/processor/library/ExportsLibrary.java
@@ -57,6 +57,7 @@ import javax.lang.model.element.VariableElement;
 import javax.lang.model.type.TypeMirror;
 
 import com.oracle.truffle.dsl.processor.ProcessorContext;
+import com.oracle.truffle.dsl.processor.expression.DSLExpression;
 import com.oracle.truffle.dsl.processor.java.ElementUtils;
 import com.oracle.truffle.dsl.processor.model.CacheExpression;
 import com.oracle.truffle.dsl.processor.model.MessageContainer;
@@ -73,6 +74,7 @@ public final class ExportsLibrary extends Template {
     private int defaultExportPriority;
     private Map<CacheExpression, String> sharedExpressions;
     private VariableElement delegationVariable;
+    private DSLExpression transitionLimit;
 
     public ExportsLibrary(ProcessorContext context, TypeElement templateType, AnnotationMirror annotation, ExportsData exports, LibraryData library, TypeMirror receiverType,
                     boolean explicitReceiver) {
@@ -220,6 +222,18 @@ public final class ExportsLibrary extends Template {
 
     public int getDefaultExportPriority() {
         return defaultExportPriority;
+    }
+
+    public void setTransitionLimit(DSLExpression allowTransition) {
+        this.transitionLimit = allowTransition;
+    }
+
+    public DSLExpression getTransitionLimit() {
+        return transitionLimit;
+    }
+
+    public boolean isAllowTransition() {
+        return transitionLimit != null;
     }
 
 }